Test unit DeliveryService - WIP
