+++ package.json
@@ -28,6 +28,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.2.0",
+    "@grpc/proto-loader": "^0.3.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
     "babel-preset-env": "^1.7.0",
